Initial commit
I created this Datapack from scratch in a Vanilla Minecraft 1.16.4 world. I decided to use Git(Hub) for version-controlling.

This initial commit contains the current state of the project, including many unpatched false positives/false negatives/bugs. It's a sneak peek of the current project progress and it's not suitable for application in production environments (e.g. Minecraft servers).

In detail, this commit contains the following features:
- Base of NoCheatMinus
- Installation routine
- Default configuration file
- Configuration file management
- UUID-system
- Appropriate check
- Appropriate fall damage sub check (calculates the amount of fall damage (depending on the fall distance and the current health) and compares it to the actual amount of taken damage)
- Appropriate speed sub check (compares several movement data in different movement states (walk, sprint, sneak) with the configurated appropriate values. This sub check contains mechanics blocking false positives (FP) related to movement state/environment transitions (LostSprint/LostWalk/LeaveIce) and player-side FP-sources (Speed potion effects with amplifier 0 and 1))